Fix order process.

+++ web-shop/main/java/ua/nure/kopaniev/cart/UserCart.java
@@ -20,7 +20,7 @@ public interface UserCart {
 
     List<Item> getItems();
 
-    long getCountOfItems(Item item);
+    int getCountOfItems(Item item);
 
     long getCount();
