+++ web-shop/main/java/ua/nure/kopaniev/cart/UserCartImpl.java
@@ -14,19 +14,19 @@ import java.util.Map;
 @Service
 @Scope(value = "session", proxyMode = ScopedProxyMode.TARGET_CLASS)
 public class UserCartImpl implements UserCart, Serializable {
-    private Map<Item, Long> items = new LinkedHashMap<>();
+    private Map<Item, Integer> items = new LinkedHashMap<>();
 
     @Override
     public void add(Item item) {
-        long count = items.getOrDefault(item, 0L);
+        int count = items.getOrDefault(item, 0);
         items.put(item, ++count);
     }
 
     @Override
     public long increase(long id) {
-        for (Map.Entry<Item, Long> entry : items.entrySet()) {
+        for (Map.Entry<Item, Integer> entry : items.entrySet()) {
             if (entry.getKey().getId() == id) {
-                Long count = 1 + entry.getValue();
+                int count = 1 + entry.getValue();
                 items.put(entry.getKey(), count);
                 return count;
             }
@@ -36,9 +36,9 @@ public class UserCartImpl implements UserCart, Serializable {
 
     @Override
     public long decrease(long id) {
-        for (Map.Entry<Item, Long> entry : items.entrySet()) {
+        for (Map.Entry<Item, Integer> entry : items.entrySet()) {
             if (entry.getKey().getId() == id) {
-                Long count = entry.getValue() - 1;
+                int count = entry.getValue() - 1;
                 if (count >= 1) {
                     entry.setValue(count);
                 }
@@ -51,7 +51,7 @@ public class UserCartImpl implements UserCart, Serializable {
     @Override
     public void remove(long id) {
         Item t = null;
-        for (Map.Entry<Item, Long> entry : items.entrySet()) {
+        for (Map.Entry<Item, Integer> entry : items.entrySet()) {
             if (entry.getKey().getId() == id) {
                 t = entry.getKey();
                 break;
@@ -71,8 +71,8 @@ public class UserCartImpl implements UserCart, Serializable {
     }
 
     @Override
-    public long getCountOfItems(Item item) {
-        return items.getOrDefault(item, 0L);
+    public int getCountOfItems(Item item) {
+        return items.getOrDefault(item, 0);
     }
 
     @Override
@@ -81,7 +81,7 @@ public class UserCartImpl implements UserCart, Serializable {
                 .values()
                 .stream()
                 .reduce((c1, c2) -> c1 + c2)
-                .orElse(0L);
+                .orElse(0);
     }
 
     @Override
